Added additional method to Student
The method getHeardLiked() returns an array of Booleans indicating whether or not the student heard or liked a song. Index 0 in the array indicates whether or not the student heard the song, and index 1 in the array indicates whether or not the student liked the song. This will be useful for collecting data from the Student object to put into the Glyph. Uses Boolean instead of boolean because several lines in the input files are missing, so it is necessary that null can be used.